pyproject.toml
[tool.poetry]
name = "test"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
asgiref = "3.8.1"
certifi = "2024.8.30"
charset-normalizer = "3.4.0"
coreapi = "2.3.3"
coreschema = "0.0.4"
django = "5.1.2"
django-environ = "0.11.2"
django-filter = "24.3"
djangorestframework = "3.15.2"
djangorestframework-simplejwt = "5.3.1"
drf-yasg = "1.21.8"
idna = "3.10"
inflection = "0.5.1"
itypes = "1.2.0"
jinja2 = "3.1.4"
markupsafe = "3.0.2"
mysqlclient = "2.2.5"
openapi-codec = "1.3.2"
packaging = "24.1"
pyjwt = "2.9.0"
python-dotenv = "1.0.1"
pytz = "2024.2"
pyyaml = "6.0.2"
requests = "2.32.3"
simplejson = "3.19.3"
sqlparse = "0.5.1"
typing-extensions = "4.12.2"
uritemplate = "4.1.1"
urllib3 = "2.2.3"
gunicorn = "^23.0.0"
mysql-connector-python = "^9.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
